separate test_pd_client
