+++ Cargo.toml
@@ -265,6 +265,7 @@ members = [
   "components/test_coprocessor",
   "components/test_coprocessor_plugin/example_plugin",
   "components/test_pd",
+  "components/test_pd_client",
   "components/test_raftstore",
   "components/test_sst_importer",
   "components/test_storage",
